set code status for the responses

+++ app/Http/Controllers/MarcaController.php
@@ -19,7 +19,7 @@ class MarcaController extends Controller
     public function index()
     {
         $marcas = $this->marca->all();
-        return $marcas;
+        return response()->json($marcas, 200);
     }
 
     /**
@@ -31,7 +31,7 @@ class MarcaController extends Controller
     public function store(Request $request)
     {
         $marca = $this->marca->create($request->all());
-        return $marca;
+        return response()->json($marca, 201);
     }
 
     /**
@@ -43,7 +43,10 @@ class MarcaController extends Controller
     public function show($id)
     {
         $marca = $this->marca->find($id);
-        return $marca;
+        if ($marca === null) {
+            return response()->json(['erro' => 'O registro solicitado não existe'], 404);
+        }
+        return response()->json($marca, 200);
     }
 
     /**
@@ -56,8 +59,13 @@ class MarcaController extends Controller
     public function update(Request $request, $id)
     {
         $marca = $this->marca->find($id);
+
+        if ($marca === null) {
+            return response()->json(['erro' => 'Não foi possível realizar a atualização. Registro inexistente'], 404);
+        }
+
         $marca->update($request->all());
-        return $marca;
+        return response()->json($marca, 200);
     }
 
     /**
@@ -69,7 +77,12 @@ class MarcaController extends Controller
     public function destroy($id)
     {
         $marca = $this->marca->find($id);
+
+        if ($marca === null) {
+            return response()->json(['erro' => 'Não foi possível realizar a exclusão. Registro inexistente'], 404);
+        }
+
         $marca->delete();
-        return ['msg' => 'A marca foi excluída com sucesso'];
+        return response()->json(['msg' => 'A marca foi excluída com sucesso'], 200);
     }
 }
